fix: add proxy support for downloading FCS keys
