+++ cmd/ipsw/cmd/fw/aea.go
@@ -50,6 +50,7 @@ func init() {
 	aeaCmd.Flags().StringP("pem", "p", "", "AEA private_key.pem file")
 	aeaCmd.Flags().String("pem-db", "", "AEA pem DB JSON file")
 	aeaCmd.Flags().BoolP("encrypt", "e", false, "AEA encrypt file")
+	aeaCmd.Flags().String("proxy", "", "HTTP/HTTPS proxy")
 	aeaCmd.Flags().Bool("insecure", false, "Allow insecure connections")
 	aeaCmd.Flags().StringP("output", "o", "", "Folder to extract files to")
 	aeaCmd.MarkFlagDirname("output")
@@ -62,6 +63,7 @@ func init() {
 	viper.BindPFlag("fw.aea.pem", aeaCmd.Flags().Lookup("pem"))
 	viper.BindPFlag("fw.aea.pem-db", aeaCmd.Flags().Lookup("pem-db"))
 	viper.BindPFlag("fw.aea.encrypt", aeaCmd.Flags().Lookup("encrypt"))
+	viper.BindPFlag("fw.aea.proxy", aeaCmd.Flags().Lookup("proxy"))
 	viper.BindPFlag("fw.aea.insecure", aeaCmd.Flags().Lookup("insecure"))
 	viper.BindPFlag("fw.aea.output", aeaCmd.Flags().Lookup("output"))
 }
@@ -84,6 +86,7 @@ var aeaCmd = &cobra.Command{
 		pemFile := viper.GetString("fw.aea.pem")
 		pemDB := viper.GetString("fw.aea.pem-db")
 		doEncrypt := viper.GetBool("fw.aea.encrypt")
+		proxy := viper.GetString("fw.aea.proxy")
 		insecure := viper.GetBool("fw.aea.insecure")
 		output := viper.GetString("fw.aea.output")
 		// validate flags
@@ -138,7 +141,7 @@ var aeaCmd = &cobra.Command{
 			if err != nil {
 				return fmt.Errorf("failed to parse AEA: %v", err)
 			}
-			pkmap, err := metadata.GetPrivateKey(nil, pemDB, false, insecure)
+			pkmap, err := metadata.GetPrivateKey(nil, pemDB, false, proxy, insecure)
 			if err != nil {
 				return fmt.Errorf("failed to get private key: %v", err)
 			}
@@ -167,7 +170,7 @@ var aeaCmd = &cobra.Command{
 			if err != nil {
 				return fmt.Errorf("failed to parse AEA: %v", err)
 			}
-			wkey, err := metadata.DecryptFCS(pemData, pemDB, insecure)
+			wkey, err := metadata.DecryptFCS(pemData, pemDB, proxy, insecure)
 			if err != nil {
 				return fmt.Errorf("failed to HPKE decrypt fcs-key: %v", err)
 			}
@@ -195,6 +198,7 @@ var aeaCmd = &cobra.Command{
 				PrivKeyData: pemData,
 				B64SymKey:   base64Key,
 				PemDB:       pemDB,
+				Proxy:  	 proxy,
 				Insecure:    insecure,
 			})
 			if err != nil {
