+++ cmd/ipsw/cmd/mdevs.go
@@ -109,6 +109,7 @@ var mdevsCmd = &cobra.Command{
 				Input:    dmgPath,
 				Output:   filepath.Dir(dmgPath),
 				PemDB:    pemDB,
+				Proxy:    "",    // TODO: make proxy configurable
 				Insecure: false, // TODO: make insecure configurable
 			})
 			if err != nil {
