+++ cmd/ipsw/cmd/sb/sb_diff.go
@@ -117,6 +117,7 @@ var sbDiffCmd = &cobra.Command{
 						Input:    dmgPath,
 						Output:   filepath.Dir(dmgPath),
 						PemDB:    pemDB,
+						Proxy:    "",    // TODO: make proxy configurable
 						Insecure: false, // TODO: make insecure configurable
 					})
 					if err != nil {
